fix(cypher): restore CALL scope CI contracts
