+++ crates/kglite/src/graph/languages/cypher/executor/write.rs
@@ -612,6 +612,7 @@ fn clause_needs_implicit_row(clause: &Clause) -> bool {
         Clause::With(_)
             | Clause::Unwind(_)
             | Clause::Call(_)
+            | Clause::CallSubquery { .. }
             | Clause::Create(_)
             | Clause::Merge(_)
             | Clause::Foreach { .. }
@@ -2450,96 +2451,5 @@ mod remove_columnar_tests;
 mod rel_constraint_tests;
 
 #[cfg(test)]
-mod is_mutation_query_tests {
-    use super::super::super::ast::*;
-    use super::is_mutation_query;
-
-    fn query(clauses: Vec<Clause>) -> CypherQuery {
-        CypherQuery {
-            clauses,
-            explain: false,
-            profile: false,
-            output_format: OutputFormat::Default,
-            optimizer_tags: Vec::new(),
-        }
-    }
-
-    fn create_clause() -> Clause {
-        Clause::Create(CreateClause {
-            patterns: Vec::new(),
-        })
-    }
-
-    fn return_clause() -> Clause {
-        Clause::Return(ReturnClause {
-            items: Vec::new(),
-            distinct: false,
-            having: None,
-            lazy_eligible: false,
-            group_limit_hint: None,
-        })
-    }
-
-    #[test]
-    fn plain_read_is_not_a_mutation() {
-        assert!(!is_mutation_query(&query(vec![return_clause()])));
-    }
-
-    #[test]
-    fn top_level_write_is_a_mutation() {
-        assert!(is_mutation_query(&query(vec![create_clause()])));
-    }
-
-    #[test]
-    fn write_inside_call_subquery_body_is_a_mutation() {
-        // CALL { CREATE (...) RETURN ... } — the body carries a write,
-        // so the enclosing query must classify as a mutation even though
-        // the outer clause is a CallSubquery (not itself a write clause).
-        let body = Box::new(query(vec![create_clause(), return_clause()]));
-        let call = Clause::CallSubquery {
-            import: CallSubqueryImport::Legacy(Vec::new()),
-            body,
-        };
-        assert!(is_mutation_query(&query(vec![call, return_clause()])));
-    }
-
-    #[test]
-    fn nested_write_inside_call_subquery_body_is_a_mutation() {
-        // CALL { CALL { CREATE (...) RETURN ... } RETURN ... } — recursion
-        // must reach an arbitrarily-deep nested body.
-        let inner = Box::new(query(vec![create_clause(), return_clause()]));
-        let inner_call = Clause::CallSubquery {
-            import: CallSubqueryImport::Legacy(Vec::new()),
-            body: inner,
-        };
-        let outer = Box::new(query(vec![inner_call, return_clause()]));
-        let outer_call = Clause::CallSubquery {
-            import: CallSubqueryImport::Legacy(Vec::new()),
-            body: outer,
-        };
-        assert!(is_mutation_query(&query(vec![outer_call])));
-    }
-
-    #[test]
-    fn read_only_call_subquery_body_is_not_a_mutation() {
-        let body = Box::new(query(vec![return_clause()]));
-        let call = Clause::CallSubquery {
-            import: CallSubqueryImport::Legacy(Vec::new()),
-            body,
-        };
-        assert!(!is_mutation_query(&query(vec![call, return_clause()])));
-    }
-
-    #[test]
-    fn write_inside_union_arm_is_a_mutation() {
-        // UNION arms also recurse — a write in either arm makes the
-        // query a mutation.
-        let arm = Box::new(query(vec![create_clause(), return_clause()]));
-        let union = Clause::Union(UnionClause {
-            all: false,
-            query: arm,
-            kind: SetOpKind::Union,
-        });
-        assert!(is_mutation_query(&query(vec![return_clause(), union])));
-    }
-}
+#[path = "write_mutation_query_tests.rs"]
+mod is_mutation_query_tests;
