+++ crates/kglite/src/graph/languages/cypher/planner/mod.rs
@@ -2,7 +2,6 @@
 
 use super::ast::*;
 use crate::datatypes::values::Value;
-use crate::graph::core::pattern_matching::PatternElement;
 use crate::graph::schema::DirGraph;
 use std::collections::{HashMap, HashSet};
 
@@ -14,6 +13,7 @@ pub mod cost_model;
 pub mod fusion;
 pub mod index_selection;
 pub mod join_order;
+mod nested;
 mod node_anchor;
 pub mod rel_predicate_pushdown;
 pub mod schema_check;
@@ -37,6 +37,7 @@ use join_order::{
     optimize_pattern_start_node, reorder_cyclic_pattern_edges, reorder_match_clauses,
     reorder_match_patterns,
 };
+pub(crate) use nested::import_pattern_anchors_in_arm;
 use node_anchor::anchor_element_id;
 use rel_predicate_pushdown::extract_pushable_rel_predicates_with_params;
 use var_length_lowering::lower_fixed_var_length_hops;
@@ -98,7 +99,10 @@ type PassFn = fn(&mut CypherQuery, &PassCtx);
 /// standing, so a hint written against a body it could not see changes
 /// no answer.
 pub const PASSES: &[(&str, PassFn)] = &[
-    ("optimize_nested_queries", pass_optimize_nested_queries),
+    (
+        "optimize_nested_queries",
+        nested::pass_optimize_nested_queries,
+    ),
     // `*k..k` → k explicit hops. Runs FIRST among the structural rewrites:
     // every pass below bails on `var_length.is_some()`, so a segment lowered
     // any later would inherit none of them.
@@ -346,184 +350,6 @@ fn optimize_with_disabled_scoped(
 /// module docs.
 fn pass_lower_fixed_var_length_hops(query: &mut CypherQuery, _ctx: &PassCtx) {
     lower_fixed_var_length_hops(query)
-}
-
-/// **Pass:** `optimize_nested_queries` — Recurse the optimizer into
-/// every nested query: UNION right-arms and `CALL { }` subquery bodies.
-/// Inherits the parent's `disabled` set so diagnostic toggles propagate
-/// to the inner planner pipeline — including the `disable_optimizer=True`
-/// expansion, which puts every pass name (this one among them) into
-/// `disabled`. When THIS pass is itself disabled the recursion never
-/// runs, so a fully-disabled optimizer leaves bodies un-optimized too,
-/// making the differential corpus's optimized-vs-naive comparison
-/// meaningful for subquery bodies.
-///
-/// This pass OWNS `CALL { }` body optimization (the executor runs the body
-/// exactly as planned here). Two body shapes are optimized differently:
-///
-/// - A body whose patterns do not anchor on an imported variable uses the
-///   full pipeline. A graph-global aggregate remains independent of the
-///   per-row seed, so the seed-ignoring fused operators are correct.
-/// - A body whose patterns anchor on an imported variable disables the
-///   [`seed_ignoring_fusion_passes`] for that body — they
-///   emit plan-time-anchored operators that ignore the per-row seed and
-///   would return the GLOBAL count for every outer row. Disabling them
-///   leaves a plain `Match`/`Return` that honours the seeded binding via
-///   CSR adjacency (§3.2). The disable is unioned with the inherited
-///   `disabled` set so an outer toggle still propagates.
-fn pass_optimize_nested_queries(query: &mut CypherQuery, ctx: &PassCtx) {
-    let mut visible = ctx.initial_scope.clone();
-    for clause in &mut query.clauses {
-        match clause {
-            Clause::Union(ref mut u) => {
-                optimize_with_disabled_scoped(
-                    &mut u.query,
-                    ctx.graph,
-                    ctx.params,
-                    ctx.disabled,
-                    ctx.initial_scope,
-                    ctx.global_scope,
-                );
-            }
-            Clause::CallSubquery {
-                ref import,
-                ref mut body,
-            } => {
-                let imports: HashSet<String> = match import {
-                    CallSubqueryImport::Legacy(names) | CallSubqueryImport::Named(names) => {
-                        names.iter().cloned().collect()
-                    }
-                    CallSubqueryImport::All => visible.clone(),
-                    CallSubqueryImport::Empty => HashSet::new(),
-                };
-                let import_names: Vec<String> = imports.iter().cloned().collect();
-                let anchors = import_pattern_anchors(body, &import_names);
-                let empty_globals = HashSet::new();
-                let body_globals = if matches!(
-                    import,
-                    CallSubqueryImport::Named(_) | CallSubqueryImport::All
-                ) {
-                    &imports
-                } else {
-                    &empty_globals
-                };
-                // Legacy set-arm importing WITH clauses are executable scope
-                // boundaries and also record which outer variables belong to
-                // each arm. Removing them would make seeded execution lose
-                // that arm's imports, so the pass-through fold is disabled
-                // throughout this one body/set tree.
-                let legacy_set = matches!(import, CallSubqueryImport::Legacy(_))
-                    && body.clauses.iter().any(|c| matches!(c, Clause::Union(_)));
-                let mut merged = ctx.disabled.clone();
-                if legacy_set {
-                    merged.insert("fold_pass_through_with".to_string());
-                }
-                if !anchors.is_empty() {
-                    // Union the seed-ignoring set with the inherited disabled
-                    // set so both per-row correctness and outer diagnostic
-                    // toggles apply to the body.
-                    merged.extend(seed_ignoring_fusion_passes().iter().cloned());
-                }
-                optimize_with_disabled_scoped(
-                    body,
-                    ctx.graph,
-                    ctx.params,
-                    &merged,
-                    &imports,
-                    body_globals,
-                );
-            }
-            _ => {}
-        }
-        crate::graph::languages::cypher::planner::simplification::advance_visible_variable_scope(
-            &mut visible,
-            clause,
-        );
-        visible.extend(ctx.global_scope.iter().cloned());
-    }
-}
-
-/// The subset of `import` names that appear as a `MATCH` / `OPTIONAL
-/// MATCH` pattern element in a correlated `CALL { }` body (so the body
-/// anchors on the seeded binding). Non-empty ⇒ the seed-ignoring fusion
-/// passes must be disabled when optimizing the body, and (in the
-/// executor) a NULL value for any of these names empties the per-row
-/// pipeline (§1.3 of the design doc).
-///
-/// Only the body's OWN clauses are scanned — a nested `CALL { }` re-binds
-/// its own imports from its own seed, so its patterns are not this body's
-/// concern.
-pub(crate) fn import_pattern_anchors(body: &CypherQuery, import: &[String]) -> Vec<String> {
-    let mut anchors: Vec<String> = Vec::new();
-    collect_import_pattern_anchors(body, import, true, &mut anchors);
-    anchors
-}
-
-/// Arm-local form used by seeded set execution. Unlike the planner-wide
-/// analysis, it stops at the current arm's set operator so a NULL import can
-/// be a scalar in one legacy arm and a pattern anchor in another.
-pub(crate) fn import_pattern_anchors_in_arm(body: &CypherQuery, import: &[String]) -> Vec<String> {
-    let mut anchors = Vec::new();
-    collect_import_pattern_anchors(body, import, false, &mut anchors);
-    anchors
-}
-
-fn collect_import_pattern_anchors(
-    body: &CypherQuery,
-    import: &[String],
-    recurse_sets: bool,
-    anchors: &mut Vec<String>,
-) {
-    for clause in &body.clauses {
-        let patterns = match clause {
-            Clause::Match(m) | Clause::OptionalMatch(m) => &m.patterns,
-            Clause::Union(set) if recurse_sets => {
-                collect_import_pattern_anchors(&set.query, import, true, anchors);
-                continue;
-            }
-            _ => continue,
-        };
-        for pattern in patterns {
-            for elem in &pattern.elements {
-                let var = match elem {
-                    PatternElement::Node(np) => np.variable.as_ref(),
-                    PatternElement::Edge(ep) => ep.variable.as_ref(),
-                };
-                if let Some(v) = var {
-                    if import.iter().any(|name| name == v) && !anchors.iter().any(|a| a == v) {
-                        anchors.push(v.clone());
-                    }
-                }
-            }
-        }
-    }
-}
-
-/// The optimizer passes that emit a graph-global / plan-time-anchored
-/// operator (`FusedCount*`, `FusedMatch*Aggregate`, `FusedNodeScan*`)
-/// which IGNORES the incoming seed row. Disabled when a correlated body
-/// anchors on an imported variable (see [`pass_optimize_nested_queries`]).
-///
-/// These names MUST stay in sync with `PASSES`; each is a registered pass
-/// name. A future `fuse_call_subquery_aggregate` pass (design §Q7) would
-/// be the correct seed-AWARE replacement and would NOT belong here.
-pub(crate) fn seed_ignoring_fusion_passes() -> &'static HashSet<String> {
-    static PASSES_SET: std::sync::OnceLock<HashSet<String>> = std::sync::OnceLock::new();
-    PASSES_SET.get_or_init(|| {
-        [
-            "fuse_anchored_edge_count",
-            "fuse_count_short_circuits",
-            "fuse_optional_match_aggregate",
-            "fuse_match_return_aggregate",
-            "fuse_match_with_aggregate",
-            "fuse_match_with_aggregate_top_k",
-            "fuse_node_scan_aggregate",
-            "fuse_node_scan_top_k",
-        ]
-        .iter()
-        .map(|s| s.to_string())
-        .collect()
-    })
 }
 
 /// **Pass:** `push_where_into_match` — Move comparison predicates from
