+++ tests/api-baselines/source-quality.json
@@ -142,9 +142,9 @@
     },
     {
       "path": "crates/kglite/src/graph/languages/cypher/executor/call_clause.rs",
-      "qualified_name": "crate::CypherExecutor<'a>::execute_call",
-      "lines": 523,
-      "branches": 101,
+      "qualified_name": "crate::CypherExecutor<'a>::execute_resolved_call_once",
+      "lines": 470,
+      "branches": 95,
       "nesting": 5
     },
     {
@@ -261,10 +261,10 @@
     },
     {
       "path": "crates/kglite/src/graph/languages/cypher/planner/schema_check/mod.rs",
-      "qualified_name": "crate::validate_scope",
-      "lines": 130,
-      "branches": 56,
-      "nesting": 6
+      "qualified_name": "crate::validate_scope_with_globals",
+      "lines": 104,
+      "branches": 55,
+      "nesting": 4
     },
     {
       "path": "crates/kglite/src/graph/languages/cypher/planner/simplification.rs",
